// src/generated/client/types/BodyMeasurement.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type BodyMeasurement = {
  /**
   * @type string, date
   */
  date: string;
  /**
   * @type number
   */
  weight_kg?: number | null;
  /**
   * @type number
   */
  lean_mass_kg?: number | null;
  /**
   * @type number
   */
  fat_percent?: number | null;
  /**
   * @type number
   */
  neck_cm?: number | null;
  /**
   * @type number
   */
  shoulder_cm?: number | null;
  /**
   * @type number
   */
  chest_cm?: number | null;
  /**
   * @type number
   */
  left_bicep_cm?: number | null;
  /**
   * @type number
   */
  right_bicep_cm?: number | null;
  /**
   * @type number
   */
  left_forearm_cm?: number | null;
  /**
   * @type number
   */
  right_forearm_cm?: number | null;
  /**
   * @type number
   */
  abdomen?: number | null;
  /**
   * @type number
   */
  waist?: number | null;
  /**
   * @type number
   */
  hips?: number | null;
  /**
   * @type number
   */
  left_thigh?: number | null;
  /**
   * @type number
   */
  right_thigh?: number | null;
  /**
   * @type number
   */
  left_calf?: number | null;
  /**
   * @type number
   */
  right_calf?: number | null;
};